+++ backend/src/main/java/com/example/backend/Entities/User.java
@@ -11,7 +11,7 @@ import java.util.Objects;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column(nullable = false)
     private  String name;
 
@@ -38,7 +38,7 @@ public class User {
     public User() {
     }
 
-    public User(Integer id, String name, String email, String password, List<Board> boards) {
+    public User(Long id, String name, String email, String password, List<Board> boards) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -55,11 +55,11 @@ public class User {
     }
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
